feat(rates): validate origin and destination exist
I am unsure whether this is the best solution, but this the one I chose to implement.
Here are some alternatives I thought about:
1. The one you're seeing: use a specific query to determine whether there were any
   matches for the specified origin/destination.
   - PROS: We go tackle the validation directly. The code is simple and similar to the regular query.
   - CONS: We are making an extra round-trip for each request
2. Include the validation statement as two extra columns in the original query
   result.
   - PROS: We avoid the extra query/round-trip
   - CONS: We enlarge the result unnecessarily with redundant data. Data transformation becomes more complex (we would have to take the first result, check it is OK, create the first record, and then create the rest and append everything).
 3. Create a stored procedure with all the logic.
   - PROS: We avoid the extra round-trip (although not the extra query, unless you follow the second approach in the procedure).
   - CONS: Data logic resides in the DB, with all its consequences (migration maintenance, testing, debugging...)

If our criteria had been to choose the most efficient approach in terms of performance, we would have had to benchmark all three approaches, but that would require a considerable effort.

PS: The original rates query now has two different hierarchies to be aligned with the new query. There does not seem to be a significant performance hit.